# docs/swagger.yaml
basePath: /
definitions:
  domain.User:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  echo.HTTPError:
    properties:
      message: {}
    type: object
host: localhost:8080
info:
  contact: {}
  description: Crime app auth provides authentication for crime-app microservices.
  title: Crime app auth
  version: "1.0"
paths:
  /users/damn:
    post:
      consumes:
      - application/json
      description: authorize user in crime-app microservices ecosystem
      produces:
      - application/json
      responses:
        default:
          description: ""
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: provides signing in operation
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      description: authorize user in crime-app microservices ecosystem
      parameters:
      - description: User's consisted of login and password
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/domain.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
        default:
          description: ""
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: provides signing in operation
      tags:
      - users
  /users/register:
    post:
      consumes:
      - application/json
      description: registers user in crime-app microservices ecosystem
      parameters:
      - description: User's consisted of login and password
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/domain.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
        default:
          description: ""
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: provides signing up operation
      tags:
      - users
swagger: "2.0"